app/oauth2.py
from jose import JWTError, jwt
from datetime import datetime, timedelta

SECRET_KEY = "b2c4879749bde467b46d054bb84989f55a1a8d7c8c2f21aeb237618e53cc0a4a"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def create_access_token(data: dict):
    to_encode = data.copy()

    expire = datetime.now() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)

    to_encode.update({"exp": expire})

    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    
    return encoded_jwt
